use server data to display game reviews
